+++ packages/contracts/src/core/application.ts
@@ -5,6 +5,7 @@ import { Logger } from '../logging'
 import { ConfigStore } from '../config'
 import { Container } from '../container'
 import { BootstrapperCtor } from './bootstrapper'
+import { ErrorHandlerCtor } from './error-handler'
 
 export interface Application extends Container {
   /**
@@ -21,6 +22,13 @@ export interface Application extends Container {
    * Returns the app version.
    */
   version(): string | undefined
+
+  /**
+   * Assign the given error handler to this application instance. The error
+   * handler is used to report errors on the default logging channel and
+   * also to create responses for requests throwing errors.
+   */
+  withErrorHandler (ErrorHandler: ErrorHandlerCtor): this
 
   /**
    * Returns the root path of the application directory.