+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [3.3.0](https://github.com/supercharge/framework/compare/v3.2.0...v3.3.0) - 2022-07-xx
 
 ### Added
+- `@supercharge/contracts`
+    - extend `Application` contract by adding the `withErrorHandler` method
 - `@supercharge/view`
     - add `registerPartial(name, content)` method: register a partial view dynamically in a service provider (for example in community packages)