+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - add `request.isAjax()` method: determine whether the request is a result of an AJAX call
     - add `request.isPjax()` method: determine whether the request is a result of an PJAX call
     - add `request.isPrefetch()` method: determine whether the request is a result of a [prefetch call](https://developer.mozilla.org/en-US/docs/Glossary/Prefetch)
+    - add `request.url()` method: returns a `URL` instance allowing you to access individual parts from the request’s URI
 
 ### Updated
 - bump dependencies