+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [3.3.0](https://github.com/supercharge/framework/compare/v3.2.0...v3.3.0) - 2022-07-xx
+## [3.3.0](https://github.com/supercharge/framework/compare/v3.2.0...v3.3.0) - 2022-07-23
 
 ### Added
 - `@supercharge/contracts`